refactored packges to make it less convoluted and easier to read
