+++ src/main/java/com/the/mild/project/db/mongo/BaseDocument.java
@@ -1,4 +1,4 @@
-package com.the.mild.project.db.mongo.documents.util;
+package com.the.mild.project.db.mongo;
 
 import org.bson.Document;
