+++ src/main/java/com/the/mild/project/db/mongo/InsertDocumentEntry.java
@@ -1,4 +1,4 @@
-package com.the.mild.project.db.mongo.documents.util;
+package com.the.mild.project.db.mongo;
 
 public interface InsertDocumentEntry {
     String key();
