+++ src/main/java/com/the/mild/project/db/mongo/exceptions/CollectionNotFoundException.java
@@ -1,4 +1,4 @@
-package com.the.mild.project.db.mongo.util.exceptions;
+package com.the.mild.project.db.mongo.exceptions;
 
 public class CollectionNotFoundException extends Exception {
     public CollectionNotFoundException(String message) {
