+++ src/main/java/com/the/mild/project/db/mongo/exceptions/DocumentSerializationException.java
@@ -1,4 +1,4 @@
-package com.the.mild.project.db.mongo.util.exceptions;
+package com.the.mild.project.db.mongo.exceptions;
 
 public class DocumentSerializationException extends Exception {
     public DocumentSerializationException(String errorMessage) {
