+++ src/main/java/com/the/mild/project/server/resources/TestResourceWithMultipleParams.java
@@ -1,8 +1,8 @@
 package com.the.mild.project.server.resources;
 
-import static com.the.mild.project.server.ResourceConfig.PathParams.PATH_PARAM_EXAMPLE_ID;
-import static com.the.mild.project.server.ResourceConfig.PathParams.PATH_PARAM_ID;
-import static com.the.mild.project.server.ResourceConfig.PATH_TEST_RESOURCE_WITH_MULTIPLE_PARAMS;
+import static com.the.mild.project.ResourceConfig.PathParams.PATH_PARAM_EXAMPLE_ID;
+import static com.the.mild.project.ResourceConfig.PathParams.PATH_PARAM_ID;
+import static com.the.mild.project.ResourceConfig.PATH_TEST_RESOURCE_WITH_MULTIPLE_PARAMS;
 
 import javax.inject.Singleton;
 import javax.ws.rs.GET;
@@ -12,7 +12,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
 import com.the.mild.project.server.jackson.MultipleParamsTest;
-import com.the.mild.project.server.jackson.util.JacksonHandler;
+import com.the.mild.project.server.jackson.JacksonHandler;
 
 @Singleton
 @Path(PATH_TEST_RESOURCE_WITH_MULTIPLE_PARAMS)
